+++ http/php/guzzle/src/GuzzleRequestAdapter.php
@@ -9,9 +9,13 @@
 namespace Microsoft\Kiota\Http;
 
 
+use Exception;
 use GuzzleHttp\Client;
 use GuzzleHttp\Psr7\Request;
+use Http\Promise\FulfilledPromise;
 use Http\Promise\Promise;
+use Http\Promise\RejectedPromise;
+use Microsoft\Kiota\Abstractions\ApiException;
 use Microsoft\Kiota\Abstractions\Authentication\AuthenticationProvider;
 use Microsoft\Kiota\Abstractions\RequestAdapter;
 use Microsoft\Kiota\Abstractions\RequestInformation;
@@ -27,6 +31,7 @@ use Microsoft\Kiota\Abstractions\Types\Time;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\StreamInterface;
+use function PHPUnit\Framework\throwException;
 
 /**
  * Class GuzzleRequestAdapter
@@ -79,19 +84,27 @@ class GuzzleRequestAdapter implements RequestAdapter
     /**
      * @inheritDoc
      */
-    public function sendAsync(RequestInformation $requestInfo, string $targetClass, ?ResponseHandler $responseHandler = null): Promise
+    public function sendAsync(RequestInformation $requestInfo, array $targetCallable, ?ResponseHandler $responseHandler = null, ?array $errorMappings = null): Promise
     {
-        $response = $this->getHttpResponseMessage($requestInfo);
-        return $response->then(
-            function (ResponseInterface $result) use ($targetClass, $responseHandler) {
-                if (!$responseHandler) {
-                    $rootNode = $this->getRootParseNode($result);
-                    if ($targetClass === StreamInterface::class || is_subclass_of(StreamInterface::class, $targetClass)) {
-                        return $result->getBody();
+        return $this->getHttpResponseMessage($requestInfo)->then(
+            function (ResponseInterface $result) use ($targetCallable, $responseHandler, $errorMappings) {
+                try {
+                    $this->throwFailedResponse($result, $errorMappings);
+
+                    if ($this->shouldReturnNull($result)) {
+                        return new FulfilledPromise(null);
                     }
-                    return $rootNode->getObjectValue($targetClass);
+                    if (!$responseHandler) {
+                        $rootNode = $this->getRootParseNode($result);
+                        if ($targetCallable[0] === StreamInterface::class || is_subclass_of($targetCallable[0],StreamInterface::class, $targetCallable[0])) {
+                            return $result->getBody();
+                        }
+                        return $rootNode->getObjectValue($targetCallable);
+                    }
+                    return $responseHandler->handleResponseAsync($result);
+                } catch (ApiException $exception){
+                    return new RejectedPromise($exception);
                 }
-                return $responseHandler->handleResponseAsync($result);
             }
         );
     }
@@ -107,53 +120,23 @@ class GuzzleRequestAdapter implements RequestAdapter
     /**
      * @inheritDoc
      */
-    public function sendCollectionAsync(RequestInformation $requestInfo, string $targetClass, ?ResponseHandler $responseHandler = null): Promise
+    public function sendCollectionAsync(RequestInformation $requestInfo, array $targetCallable, ?ResponseHandler $responseHandler = null, ?array $errorMappings = null): Promise
     {
-        $response = $this->getHttpResponseMessage($requestInfo);
-        return $response->then(
-            function (ResponseInterface $result) use ($targetClass, $responseHandler) {
-                if (!$responseHandler) {
-                    $rootNode = $this->getRootParseNode($result);
-                    return $rootNode->getCollectionOfObjectValues($targetClass);
-                }
-                return $responseHandler->handleResponseAsync($result);
-            }
-        );
-    }
+        return $this->getHttpResponseMessage($requestInfo)->then(
+            function (ResponseInterface $result) use ($targetCallable, $responseHandler, $errorMappings) {
+                try {
+                    $this->throwFailedResponse($result, $errorMappings);
 
-    /**
-     * @inheritDoc
-     */
-    public function sendPrimitiveAsync(RequestInformation $requestInfo, string $primitiveType, ?ResponseHandler $responseHandler = null): Promise
-    {
-        $response = $this->getHttpResponseMessage($requestInfo);
-        return $response->then(
-            function (ResponseInterface $result) use ($primitiveType, $responseHandler) {
-                if (!$responseHandler) {
-                    $rootParseNode = $this->getRootParseNode($result);
-                    switch (strtolower($primitiveType)) {
-                        case 'int':
-                        case 'long':
-                            return $rootParseNode->getIntegerValue();
-                        case 'float':
-                            return $rootParseNode->getFloatValue();
-                        case 'bool':
-                            return $rootParseNode->getBooleanValue();
-                        case 'string':
-                            return $rootParseNode->getStringValue();
-                        case \DateTime::class:
-                            return $rootParseNode->getDateTimeValue();
-                        case \DateInterval::class:
-                            return $rootParseNode->getDateIntervalValue();
-                        case Date::class:
-                            return $rootParseNode->getDateValue();
-                        case Time::class:
-                            return $rootParseNode->getTimeValue();
-                        default:
-                            throw new \InvalidArgumentException("Unsupported primitive type $primitiveType");
+                    if ($this->shouldReturnNull($result)) {
+                        return new FulfilledPromise(null);
                     }
+                    if (!$responseHandler) {
+                        return $this->getRootParseNode($result)->getCollectionOfObjectValues($targetCallable);
+                    }
+                    return $responseHandler->handleResponseAsync($result);
+                } catch (ApiException $apiException) {
+                    return new RejectedPromise($apiException);
                 }
-                return $responseHandler->handleResponseAsync($result);
             }
         );
     }
@@ -161,16 +144,44 @@ class GuzzleRequestAdapter implements RequestAdapter
     /**
      * @inheritDoc
      */
-    public function sendPrimitiveCollectionAsync(RequestInformation $requestInfo, string $primitiveType, ?ResponseHandler $responseHandler = null): Promise
+    public function sendPrimitiveAsync(RequestInformation $requestInfo, string $primitiveType, ?ResponseHandler $responseHandler = null, ?array $errorMappings = null): Promise
     {
-        $response = $this->getHttpResponseMessage($requestInfo);
-        return $response->then(
-            function (ResponseInterface $result) use ($primitiveType, $responseHandler) {
-                if (!$responseHandler) {
-                    $rootParseNode = $this->getRootParseNode($result);
-                    return $rootParseNode->getCollectionOfPrimitiveValues($primitiveType);
+        return $this->getHttpResponseMessage($requestInfo)->then(
+            function (ResponseInterface $result) use ($primitiveType, $responseHandler, $errorMappings) {
+                try {
+                    $this->throwFailedResponse($result, $errorMappings);
+
+                    if ($this->shouldReturnNull($result)) {
+                        return new FulfilledPromise(null);
+                    }
+                    if (!$responseHandler) {
+                        $rootParseNode = $this->getRootParseNode($result);
+                        switch (strtolower($primitiveType)) {
+                            case 'int':
+                            case 'long':
+                                return $rootParseNode->getIntegerValue();
+                            case 'float':
+                                return $rootParseNode->getFloatValue();
+                            case 'bool':
+                                return $rootParseNode->getBooleanValue();
+                            case 'string':
+                                return $rootParseNode->getStringValue();
+                            case \DateTime::class:
+                                return $rootParseNode->getDateTimeValue();
+                            case \DateInterval::class:
+                                return $rootParseNode->getDateIntervalValue();
+                            case Date::class:
+                                return $rootParseNode->getDateValue();
+                            case Time::class:
+                                return $rootParseNode->getTimeValue();
+                            default:
+                                throw new \InvalidArgumentException("Unsupported primitive type $primitiveType");
+                        }
+                    }
+                    return $responseHandler->handleResponseAsync($result);
+                } catch(ApiException $apiException) {
+                    return new RejectedPromise($apiException);
                 }
-                return $responseHandler->handleResponseAsync($result);
             }
         );
     }
@@ -178,13 +189,45 @@ class GuzzleRequestAdapter implements RequestAdapter
     /**
      * @inheritDoc
      */
-    public function sendNoContentAsync(RequestInformation $requestInfo, ?ResponseHandler $responseHandler = null): Promise
+    public function sendPrimitiveCollectionAsync(RequestInformation $requestInfo, string $primitiveType, ?ResponseHandler $responseHandler = null, ?array $errorMappings = null): Promise
     {
-        $response = $this->getHttpResponseMessage($requestInfo);
-        return $response->then(
-            function (ResponseInterface $result) use ($responseHandler) {
-                if ($responseHandler) {
-                    $responseHandler->handleResponseAsync($result);
+        return $this->getHttpResponseMessage($requestInfo)->then(
+            function (ResponseInterface $result) use ($primitiveType, $responseHandler, $errorMappings) {
+                try {
+                    $this->throwFailedResponse($result, $errorMappings);
+
+                    if ($this->shouldReturnNull($result)) {
+                        return new FulfilledPromise(null);
+                    }
+                    if (!$responseHandler) {
+                        return $this->getRootParseNode($result)->getCollectionOfPrimitiveValues($primitiveType);
+                    }
+                    return $responseHandler->handleResponseAsync($result);
+                } catch (ApiException $apiException) {
+                    return new RejectedPromise($apiException);
+                }
+            }
+        );
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function sendNoContentAsync(RequestInformation $requestInfo, ?ResponseHandler $responseHandler = null, ?array $errorMappings = null): Promise
+    {
+        return $this->getHttpResponseMessage($requestInfo)->then(
+            function (ResponseInterface $result) use ($responseHandler, $errorMappings) {
+                try {
+                    $this->throwFailedResponse($result, $errorMappings);
+                    if ($this->shouldReturnNull($result)) {
+                        return new FulfilledPromise(null);
+                    }
+                    if ($responseHandler) {
+                        return $responseHandler->handleResponseAsync($result);
+                    }
+                    return null;
+                } catch (ApiException $apiException) {
+                    return new RejectedPromise($apiException);
                 }
             }
         );
@@ -265,5 +308,44 @@ class GuzzleRequestAdapter implements RequestAdapter
                 return $this->guzzleClient->send($psrRequest, $requestInformation->getRequestOptions());
             }
         );
+    }
+
+    /**
+     * @param ResponseInterface $response
+     * @param array<string, array{string, string}>|null $errorMappings
+     * @throws ApiException
+     */
+    private function throwFailedResponse(ResponseInterface $response, ?array $errorMappings): void {
+        $statusCode = $response->getStatusCode();
+        if ($statusCode >= 200 && $statusCode < 400) {
+            return;
+        }
+        $statusCodeAsString = (string)$statusCode;
+        if ($errorMappings === null || (!isset($errorMappings[$statusCodeAsString]) &&
+            !($statusCode >= 400 && $statusCode < 500 && isset($errorMappings['4XX'])) &&
+            !($statusCode >= 500 && $statusCode < 600 && isset($errorMappings["5XX"])))) {
+            throw new ApiException("the server returned an unexpected status code and no error class is registered for this code " . $statusCode);
+        }
+        /** @var array{string,string}|null $errorClass */
+        $errorClass = $errorMappings[$statusCodeAsString] ?? ($errorMappings[$statusCodeAsString[0] . 'XX'] ?? null);
+
+        try {
+            $rootParseNode = $this->getRootParseNode($response);
+            $error = $rootParseNode->getObjectValue($errorClass);
+            if (is_subclass_of($error, ApiException::class)) {
+                throw $error;
+            }
+            throw new ApiException("Unsupported error type ". get_debug_type($error));
+        } catch (\RuntimeException $exception){
+            throw new \RuntimeException("", 0, $exception);
+        }
+    }
+
+    /**
+     * @param ResponseInterface $response
+     * @return bool
+     */
+    private function shouldReturnNull(ResponseInterface $response): bool{
+        return $response->getStatusCode() === 204;
     }
 }