+++ serialization/php/json/src/JsonParseNode.php
@@ -79,7 +79,7 @@ class JsonParseNode implements ParseNode
      * @return array<Parsable|null>|null
      * @throws Exception
      */
-    public function getCollectionOfObjectValues(string $type): ?array {
+    public function getCollectionOfObjectValues(array $type): ?array {
         if ($this->jsonNode === null) {
             return null;
         }
@@ -94,15 +94,18 @@ class JsonParseNode implements ParseNode
      * @inheritDoc
      * @throws Exception
      */
-    public function getObjectValue(string $type): ?Parsable {
+    public function getObjectValue(array $type): ?Parsable {
         if ($this->jsonNode === null) {
             return null;
         }
-        if (!is_subclass_of($type, Parsable::class)){
-            throw new InvalidArgumentException("Invalid type $type provided.");
+        if (!is_subclass_of($type[0], Parsable::class)){
+            throw new InvalidArgumentException("Invalid type $type[0] provided.");
+        }
+        if (!is_callable($type, true, $callableString)) {
+            throw new \RuntimeException('Undefined method '. $type[1]);
         }
         /** @var Parsable $result */
-        $result = new $type();
+        $result = $callableString($this);
         if($this->onBeforeAssignFieldValues !== null) {
             $this->onBeforeAssignFieldValues($result);
         }
@@ -132,7 +135,7 @@ class JsonParseNode implements ParseNode
             $deserializer = $fieldDeserializers[$key] ?? null;
 
             if ($deserializer !== null){
-                $deserializer($result, new JsonParseNode($value));
+                $deserializer(new JsonParseNode($value));
             } else {
                 $key = (string)$key;
                 $additionalData[$key] = $value;
@@ -155,6 +158,20 @@ class JsonParseNode implements ParseNode
             throw new InvalidArgumentException('Invalid enum provided.');
         }
         return new $targetEnum($this->jsonNode);
+    }
+
+    /**
+     * @return array<Enum|null>|null
+     */
+    public function getCollectionOfEnumValues(string $targetClass): ?array {
+        if ($this->jsonNode === null) {
+            return null;
+        }
+        return array_map(static function ($val) use($targetClass) {
+            return $val->getEnumValue($targetClass);
+        }, array_map(static function ($value) {
+            return new JsonParseNode($value);
+        }, $this->jsonNode));
     }
 
     /**
@@ -226,9 +243,6 @@ class JsonParseNode implements ParseNode
             default:
                 if (is_subclass_of($type, Enum::class)){
                     return $this->getEnumValue($type);
-                }
-                if (is_subclass_of($type, Parsable::class)){
-                    return $this->getObjectValue($type);
                 }
                 if (is_subclass_of($type, StreamInterface::class)) {
                     return $this->getBinaryContent();